fix($intercept-submit): ensure invalid fields are highlighted by browser
The previous approach to intercepting a call to the form's native
submit function was missing one important piece - invalid field highlting.
Simply triggering the submit event will not highlight the invalid field
(if the browser naturally supports highlighting invalid form fields on a
failed submit).  To force the browser to do this programmatically, we must
click a hidden submit button if the form in invalid.  Otherwise, trigger
the submit event.

ref #15